+++ CHANGELOG.md
@@ -4,11 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.1.0] - 2020-03-03
+### Added
+- Wrappers for `Intrinsics` and `Extrinsics`
+  - Acquisition of `Intrinsics` via `StreamProfile::get_intrinsics()`
+  - Acquisition of `Extrinsics` via `StreamProfile::get_extrinsics()` or `StreamProfile::get_extrinsics_to()`
+
+## [1.0.1] - 2019-12-12
 ### Added
 - Add `Pipeline::poll_for_frames()`
 - Add `Playback` and the functionality for setting rosbag playback to non real-time
-
 
 ## [1.0.0] - 2019-11-11
 ### Added